# docs/mobile-openapi.yaml
openapi: 3.0.3
info:
  title: NGA Admin Mobile API (subset)
  version: "1.0.0"
  description: |
    Staff native clients use Bearer JWT on the same Next.js `/api/*` hosts as the web app.
    Optional header `X-NGA-Client-Version` identifies app builds for server logs.
servers:
  - url: https://{host}
    variables:
      host:
        default: example.com
...
paths:
  /api/auth:
    post:
      tags: [auth]
      summary: Login or other auth actions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, email, password]
              properties:
                action: { type: string, example: login }
                email: { type: string, format: email }
                password: { type: string }
                client: { type: string, enum: [native, android] }
      responses:
        "200":
          description: Success; staff may include `tokens` when client is native
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenBundle:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, example: 3600 }
    LoginResponse:
      type: object
      properties:
        success: { type: boolean }
        data: { type: object }
        tokens:
          $ref: "#/components/schemas/TokenBundle"
        message: { type: string }
...